Make it buildable under Visual studio

+++ program.cpp
@@ -134,7 +134,7 @@ namespace affine_ciphers_ns {
         std::mt19937 mt(rd());
         std::uniform_int_distribution<std::size_t> a_dist(0
                 , (m_settings.text_lang == settings::Eng ? possible_a_eng.size() : possible_a_rus.size()));
-        std::uniform_int_distribution<std::uint8_t > b_dist(1
+        std::uniform_int_distribution<std::uint32_t > b_dist(1
                 , (m_settings.text_lang == settings::Eng ? eng_dict.size() + 1 : rus_dict.size() + 1));
 
         const auto a_idx = a_dist(mt);
